Relocated CSS injection into page's JS
