+++ lib/katex-screenshot.js
@@ -44,28 +44,11 @@ program.action(function main (infile, outfile) {
     // Load our requested file
     let inputKatexStr = fs.readFileSync(infile, 'utf8');
 
-    // Load our KaTeX CSS and inline its fonts
-    let katexCssContent = fs.readFileSync(require.resolve('katex/dist/katex.css'), 'utf8');
-    katexCssContent = katexCssContent.replace(/url\(["'](fonts\/[^\)]+)["']\)/g,
-        function handleFontsReplacement (urlStr, urlStrContent, urlStrIndex) {
-      // urlStr = url('fonts/KaTeX_Size1-Regular.eot')
-      // urlStrContent = fonts/KaTeX_Size1-Regular.eot
-      // urlStrIndex = 3107
-      assert(!urlStrContent.includes('..'), `Found .. in ${urlStrContent}`);
-      // Strip away `iefix` suffixes
-      urlStrContent = urlStrContent.replace(/#iefix$/, '');
-      let urlFilepath = require.resolve(`katex/dist/${urlStrContent}`);
-      return `url('file://${urlFilepath}')`;
-    });
-
     // Parse our KaTeX and built it into our HTML
     let htmlTemplateStr = fs.readFileSync(__dirname + '/renderer/index.html', 'utf8');
     let inputHtmlStr = katex.renderToString(inputKatexStr);
     assert(htmlTemplateStr.includes('{{input_html}}'));
-    assert(htmlTemplateStr.includes('<!-- katex_css_html -->'));
-    let htmlStr = htmlTemplateStr
-      .replace('{{input_html}}', inputHtmlStr)
-      .replace('<!-- katex_css_html -->', `<style>${katexCssContent}</style>`);
+    let htmlStr = htmlTemplateStr.replace('{{input_html}}', inputHtmlStr);
 
     // Generate our browser window
     let browserWindow = new BrowserWindow({
