+++ lib/renderer/index.html
@@ -17,7 +17,7 @@
         padding: 5px;
       }
     </style>
-    <!-- katex_css_html -->
+    <!-- KaTeX CSS injected via `index.js` -->
   </head>
   <body>
     <div id="container">{{input_html}}</div>
