+++ lib/renderer/index.js
@@ -1,5 +1,6 @@
 // Load in our dependencies
 const assert = require('assert');
+const fs = require('fs');
 
 const remote = require('electron').remote;
 const ipcRenderer = require('electron').ipcRenderer;
@@ -7,6 +8,26 @@ const ipcRenderer = require('electron').ipcRenderer;
 // Add an error handler
 window.addEventListener('error', function handleError (err) {
   ipcRenderer.sendSync('browser-error', err.message);
+});
+
+// Inject our own CSS into the page
+// DEV: We inject our CSS within the document as it's too bulky for data URIs
+let katexCssContent = fs.readFileSync(require.resolve('katex/dist/katex.css'), 'utf8');
+katexCssContent = katexCssContent.replace(/url\(["'](fonts\/[^\)]+)["']\)/g,
+    function handleFontsReplacement (urlStr, urlStrContent, urlStrIndex) {
+  // urlStr = url('fonts/KaTeX_Size1-Regular.eot')
+  // urlStrContent = fonts/KaTeX_Size1-Regular.eot
+  // urlStrIndex = 3107
+  assert(!urlStrContent.includes('..'), `Found .. in ${urlStrContent}`);
+  // Strip away `iefix` suffixes
+  urlStrContent = urlStrContent.replace(/#iefix$/, '');
+  let fontContentBase64 = fs.readFileSync(require.resolve(`katex/dist/${urlStrContent}`), 'base64');
+  return 'url(data:text/plain;base64,' + fontContentBase64 + ')';
+});
+window.addEventListener('DOMContentLoaded', function handleDOMContentLoaded () {
+  let katexStyleEl = document.createElement('style');
+  katexStyleEl.innerHTML = katexCssContent;
+  document.head.appendChild(katexStyleEl);
 });
 
 // When our content loads
